+++ waldreg/application-layer/reward-controller/src/main/java/org/waldreg/controller/reward/RewardController.java
@@ -1,14 +1,21 @@
 package org.waldreg.controller.reward;
 
+import java.util.List;
+import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import org.wadlreg.reward.tag.RewardTagManager;
+import org.wadlreg.reward.tag.dto.RewardTagDto;
 import org.wadlreg.reward.users.UsersRewardManager;
+import org.waldreg.character.aop.annotation.PermissionVerifying;
 import org.waldreg.controller.reward.mapper.ControllerRewardTagMapper;
 import org.waldreg.controller.reward.request.RewardTagRequest;
+import org.waldreg.controller.reward.response.RewardTagResponse;
+import org.waldreg.token.aop.annotation.Authenticating;
 
 @RestController
 public class RewardController{
@@ -26,11 +33,21 @@ public class RewardController{
         this.controllerRewardTagMapper = controllerRewardTagMapper;
     }
 
+    @Authenticating
+    @PermissionVerifying("Reward manager")
     @PostMapping("/reward-tag")
     public void createRewardTag(@RequestBody @Validated RewardTagRequest rewardTagRequest){
         rewardTagManager.createRewardTag(
                 controllerRewardTagMapper.rewardTagRequestToRewardTagDto(rewardTagRequest)
         );
+    }
+
+    @Authenticating
+    @PermissionVerifying("Reward manager")
+    @GetMapping("/reward-tag")
+    public Map<String, List<RewardTagResponse>> readRewardTagMap(){
+        List<RewardTagDto> rewardTagDtoList = rewardTagManager.readRewardTagList();
+        return Map.of("reward_tags", controllerRewardTagMapper.rewardTagDtoListToRewardTagResponseList(rewardTagDtoList));
     }
 
 }
